refactor: move buffer utilities into common.ts
